#3 Feat: 헤더에 드롭다운 메뉴 기능 구현 및 헤더 컴포넌트 내부 모듈화 작업
