+++ contexts/AuthProvider.tsx
@@ -84,9 +84,8 @@ export default function AuthProvider({ children }: AuthProviderProps) {
     setUser(null);
   };
 
-  const updateUser = async () => {
+  const initUser = async () => {
     const userInfoResponse = await getUser();
-    console.log('userInfoResponse:', userInfoResponse);
     if (userInfoResponse) {
       const { id, name, profile } = userInfoResponse;
 
@@ -105,7 +104,7 @@ export default function AuthProvider({ children }: AuthProviderProps) {
   useEffect(() => {
     const { accessToken } = getTokens();
     if (accessToken) {
-      updateUser();
+      initUser();
       setIsLoggedIn(true);
     }
   }, [isLoggedIn]);
