+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import Header from '@/components/@shared/Header';
+import Header from '@/components/@shared/Header/Header';
 import Providers from '@/contexts/Providers';
 import '@/styles/globals.scss';
 import type { AppProps } from 'next/app';
